move layout to packages
