+++ packages/@guardian/src-layout/Columns/Columns.stories.tsx
@@ -1,9 +1,6 @@
 import React, { HTMLAttributes } from 'react';
-import { Story } from '../../../../../lib/@types/storybook-emotion-10-fixes';
-import {
-	asChromaticStory,
-	asPlayground,
-} from '../../../../../lib/story-intents';
+import { Story } from '../../../../lib/@types/storybook-emotion-10-fixes';
+import { asChromaticStory, asPlayground } from '../../../../lib/story-intents';
 import { Container } from '../Container/Container';
 import { Column } from './Column';
 import { Columns, ColumnsProps } from './Columns';
