+++ packages/@guardian/src-layout/Container/Container.stories.tsx
@@ -1,9 +1,6 @@
 import { Container, ContainerProps } from './Container';
-import {
-	asChromaticStory,
-	asPlayground,
-} from '../../../../../lib/story-intents';
-import { Story } from '../../../../../lib/@types/storybook-emotion-10-fixes';
+import { asChromaticStory, asPlayground } from '../../../../lib/story-intents';
+import { Story } from '../../../../lib/@types/storybook-emotion-10-fixes';
 
 export default {
 	title: 'Source/src-layout/Container',
