+++ packages/@guardian/src-layout/Stack/Stack.stories.tsx
@@ -1,26 +1,18 @@
-import React from 'react';
-import { Story } from '../../../../../lib/@types/storybook-emotion-10-fixes';
-import { Inline, InlineProps } from './Inline';
-import {
-	asChromaticStory,
-	asPlayground,
-} from '../../../../../lib/story-intents';
+import { Stack, StackProps } from './Stack';
+import { asChromaticStory, asPlayground } from '../../../../lib/story-intents';
+import { Story } from '../../../../lib/@types/storybook-emotion-10-fixes';
 
 export default {
-	title: 'Source/src-layout/Inline',
-	component: Inline,
+	title: 'Source/src-layout/Stack',
+	component: Stack,
 };
 
-const Template: Story = (args: InlineProps) => (
-	<Inline {...args}>
-		{args.children ?? (
-			<>
-				<div>[Item 1]</div>
-				<div>[Item 2]</div>
-				<div>[Item 3]</div>
-			</>
-		)}
-	</Inline>
+const Template: Story = (args: StackProps) => (
+	<Stack {...args}>
+		<div>Item 1</div>
+		<div>Item 2</div>
+		<div>Item 3</div>
+	</Stack>
 );
 
 // *****************************************************************************
@@ -30,8 +22,8 @@ asPlayground(Playground);
 
 // *****************************************************************************
 
-export const NoSpace = Template.bind({});
-asChromaticStory(NoSpace);
+export const Default = Template.bind({});
+asChromaticStory(Default);
 
 // *****************************************************************************
 
@@ -104,32 +96,3 @@ Space24.args = {
 	space: 24,
 };
 asChromaticStory(Space24);
-
-// *****************************************************************************
-
-export const LotsOfItems = Template.bind({});
-LotsOfItems.args = {
-	children: [
-		<div key={1}>[Item 1]</div>,
-		<div key={2}>[Item 2]</div>,
-		<div key={3}>[Item 3]</div>,
-		<div key={4}>[Item 4]</div>,
-		<div key={5}>[Item 5]</div>,
-		<div key={6}>[Item 6]</div>,
-		<div key={7}>[Item 7]</div>,
-		<div key={8}>[Item 8]</div>,
-		<div key={9}>[Item 9]</div>,
-		<div key={10}>[Item 10]</div>,
-		<div key={11}>[Item 11]</div>,
-		<div key={12}>[Item 12]</div>,
-		<div key={13}>[Item 13]</div>,
-		<div key={14}>[Item 14]</div>,
-		<div key={15}>[Item 15]</div>,
-		<div key={16}>[Item 16]</div>,
-		<div key={17}>[Item 17]</div>,
-		<div key={18}>[Item 18]</div>,
-		<div key={19}>[Item 19]</div>,
-		<div key={20}>[Item 20]</div>,
-	],
-};
-asChromaticStory(LotsOfItems);
