+++ packages/@guardian/src-layout/Tiles/Tiles.stories.tsx
@@ -1,8 +1,5 @@
-import { Story } from '../../../../../lib/@types/storybook-emotion-10-fixes';
-import {
-	asChromaticStory,
-	asPlayground,
-} from '../../../../../lib/story-intents';
+import { Story } from '../../../../lib/@types/storybook-emotion-10-fixes';
+import { asChromaticStory, asPlayground } from '../../../../lib/story-intents';
 import { Tiles, TilesProps } from '../index';
 
 export default {
