+++ packages/@guardian/src-layout/tsconfig.json
@@ -1,19 +1,19 @@
 {
-  "extends": "../../../../tsconfig.json",
+  "extends": "../../../tsconfig.json",
   "compilerOptions": {
     "outDir": "dist/types",
     "declaration": true,
     "emitDeclarationOnly": true,
     "composite": true
   },
-  "include": [".", "../../foundations/types/themes.d.ts"],
+  "include": [".", "../../../src/core/foundations/types/themes.d.ts"],
   "exclude": ["dist", "**/*.stories.tsx"],
   "references": [
     {
-      "path": "../../helpers"
+      "path": "../../../src/core/helpers"
     },
     {
-      "path": "../../foundations"
+      "path": "../../../src/core/foundations"
     }
   ]
 }
